move arrays and views to collections
